refactor: 팔로우 관련 로직 분리 및 import 구문 수정
